#176: add UTF-8 HttpRequestUtil tests
